Flask App/requirements.txt
certifi==2019.3.9
numpy==1.16.4
pandas==0.24.2
patsy==0.5.1
python-dateutil==2.8.0
pytz==2019.1
simplejson==3.16.0
six==1.12.0
statsmodels==0.9.0
